refactor: Replace Timer.schedule lambda with TimerTask object for consistency

+++ src/main/kotlin/de/andrena/codingaider/outputview/MarkdownDialog.kt
@@ -338,7 +338,7 @@ class MarkdownDialog(
                                                         markdownViewer.component.revalidate()
                                                     }
                                                 }
-                                            }, 300 * contentUpdateAttempts)
+                                            }, (300 * contentUpdateAttempts).toLong())
                                         }
                                     }
                                 }
